+++ example/calculator_builder.dart
@@ -17,7 +17,7 @@ Future<void> main(List<String> args) async {
 
 const __header = '''
 void main() {
-  final source = '1 + 2 * 3 * (1 + 2.0)';
+  const source = '1 + 2 * 3 * (1 + 2.0)';
   final result = parse(source);
   print(result);
 }