+++ tool/build_json_number_parser.dart
@@ -14,7 +14,7 @@ const parser = Number();
 
 const __header = r'''
 void main() {
-    final s = '100000.00123e3';
+    const s = '100000.00123e3';
     final x = double.parse(s);
     print(s);
     print(x);